fix(gui): close the #758 review findings
The stand's neck joins the theme so the one-copy claim is true
(its old inline 44 collided with the foot floor by
coincidence); the glow needle also pins the compositingGroup-
shadow pairing, without which the halo regresses to
per-primitive; the wiring suite splits before the 350 ceiling
(MonitorsChromeWiringTests); the stand doc names
MonitorArrangement.minimumCard instead of restating derived
numbers; the theme states its area-metric boundary.

docs: ui-patterns exempts the main card's accent bloom from
the shadow/halo vocabulary and drops a stale Color.accentColor
mention; design-decisions' chips paragraph updates to the
shipped two-route ruling and carries the corner-badge rule.
