+++ docs/design-decisions.md
@@ -4190,13 +4190,22 @@ than a limit of the identity.
 (pin, arrow) rather than border styles alone (accessibility), and
 automatic is drawn as an outline rather than a dimmed capsule —
 dimming is this app's inert vocabulary, and an automatic chip is
-the one most worth dragging. The chip is a menu CONTROL, and
-that is what makes its menu reachable: the same items were
-available as a context menu for a year, on a plain stack with a
-drag gesture, which cannot take focus — so the route billed as
-the keyboard and VoiceOver fallback existed only for people using
-a mouse, the one group it was not for. Right-click still opens
-it; the control is what added the other two.
+the one most worth dragging. Two routes into a chip's menu —
+drag, or right-click — and the keyboard route is knowingly
+absent: a whole-chip `Menu` made it real for one release and
+consumed the mouse-down that `.draggable` needs, so a restored
+keyboard route must not reach for the control shape that took
+the drag (the argument lives on `SpaceAssignmentChip`). The
+clear affordance never participates in the chip's layout: a
+pinned chip and an automatic chip measure identically — the ⓧ
+rides the trailing-top corner as an overlay, and hover may
+change only its tint, never its presence or any metric —
+because the chips are sized by a flow layout whose arithmetic
+(`MonitorCardChips.minChipWidth`) must hold for both states,
+and both a hover-revealed button and an in-flow trailing slot
+have shipped and died of that measurement. Decoration may ride
+the accent (the main card's bloom); the answer never rides hue
+alone (the "main" badge).
 
 ### App rules
